added new types

+++ entities/audio.go
@@ -2,7 +2,7 @@ package entities
 
 type Audio struct {
 	Type                string `json:"@type"`
-	Duration            int    `json:"duration"`
+	Duration            int64  `json:"duration"`
 	Title               string `json:"title"`
 	Performer           string `json:"performer"`
 	FileName            string `json:"file_name"`
@@ -12,13 +12,13 @@ type Audio struct {
 		Format struct {
 			Type string `json:"@type"`
 		} `json:"format"`
-		Width  int `json:"width"`
-		Height int `json:"height"`
+		Width  int64 `json:"width"`
+		Height int64 `json:"height"`
 		File   struct {
 			Type         string `json:"@type"`
-			Id           int    `json:"id"`
-			Size         int    `json:"size"`
-			ExpectedSize int    `json:"expected_size"`
+			Id           int64  `json:"id"`
+			Size         int64  `json:"size"`
+			ExpectedSize int64  `json:"expected_size"`
 			Local        struct {
 				Type                   string `json:"@type"`
 				Path                   string `json:"path"`
@@ -26,9 +26,9 @@ type Audio struct {
 				CanBeDeleted           bool   `json:"can_be_deleted"`
 				IsDownloadingActive    bool   `json:"is_downloading_active"`
 				IsDownloadingCompleted bool   `json:"is_downloading_completed"`
-				DownloadOffset         int    `json:"download_offset"`
-				DownloadedPrefixSize   int    `json:"downloaded_prefix_size"`
-				DownloadedSize         int    `json:"downloaded_size"`
+				DownloadOffset         int64  `json:"download_offset"`
+				DownloadedPrefixSize   int64  `json:"downloaded_prefix_size"`
+				DownloadedSize         int64  `json:"downloaded_size"`
 			} `json:"local"`
 			Remote struct {
 				Type                 string `json:"@type"`
@@ -36,15 +36,15 @@ type Audio struct {
 				UniqueId             string `json:"unique_id"`
 				IsUploadingActive    bool   `json:"is_uploading_active"`
 				IsUploadingCompleted bool   `json:"is_uploading_completed"`
-				UploadedSize         int    `json:"uploaded_size"`
+				UploadedSize         int64  `json:"uploaded_size"`
 			} `json:"remote"`
 		} `json:"file"`
 	} `json:"external_album_covers"`
 	Audio struct {
 		Type         string `json:"@type"`
-		Id           int    `json:"id"`
-		Size         int    `json:"size"`
-		ExpectedSize int    `json:"expected_size"`
+		Id           int64  `json:"id"`
+		Size         int64  `json:"size"`
+		ExpectedSize int64  `json:"expected_size"`
 		Local        struct {
 			Type                   string `json:"@type"`
 			Path                   string `json:"path"`
@@ -52,9 +52,9 @@ type Audio struct {
 			CanBeDeleted           bool   `json:"can_be_deleted"`
 			IsDownloadingActive    bool   `json:"is_downloading_active"`
 			IsDownloadingCompleted bool   `json:"is_downloading_completed"`
-			DownloadOffset         int    `json:"download_offset"`
-			DownloadedPrefixSize   int    `json:"downloaded_prefix_size"`
-			DownloadedSize         int    `json:"downloaded_size"`
+			DownloadOffset         int64  `json:"download_offset"`
+			DownloadedPrefixSize   int64  `json:"downloaded_prefix_size"`
+			DownloadedSize         int64  `json:"downloaded_size"`
 		} `json:"local"`
 		Remote struct {
 			Type                 string `json:"@type"`
@@ -62,7 +62,7 @@ type Audio struct {
 			UniqueId             string `json:"unique_id"`
 			IsUploadingActive    bool   `json:"is_uploading_active"`
 			IsUploadingCompleted bool   `json:"is_uploading_completed"`
-			UploadedSize         int    `json:"uploaded_size"`
+			UploadedSize         int64  `json:"uploaded_size"`
 		} `json:"remote"`
 	} `json:"audio"`
 }
